remove cloak.c and move code to keymap
